Separate Verilog synthesis settings for Petrify and MPSat
Closes #492.
